add full crud for posts

+++ backend/app/Http/Controllers/Api/V1/PostController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers\Api\V1;
 
 use App\Http\Requests\PostRequest;
 use App\Http\Resources\PostResource;
+use App\Models\Category;
 use App\Models\Post;
+use Symfony\Component\HttpFoundation\Response as ResponseAlias;
 
 class PostController extends Controller
 {
@@ -21,7 +23,19 @@ class PostController extends Controller
      */
     public function store(PostRequest $request)
     {
-        $post = Post::create($request->validated());
+        $fields = $request->validated();
+
+        $post = Post::create([
+            'user_id' => auth()->user()->getAuthIdentifier(),
+            'body' => $fields['body'],
+            'description' => $fields['description'],
+        ]);
+
+        $categories = Category::whereIn('id', $fields['categories'])->get();
+
+        foreach ($categories as $category) {
+            $category->posts()->save($post);
+        }
 
         return new PostResource($post);
     }
@@ -39,9 +53,25 @@ class PostController extends Controller
      */
     public function update(PostRequest $request, Post $post)
     {
-        $post->update($request->validated());
+        if (auth()->user()->getAuthIdentifier() == $post->user_id) {
+            $fields = $request->validated();
 
-        return new PostResource($post);
+            $post->update([
+                'body' => $fields['body'],
+                'description' => $fields['description'],
+            ]);
+
+            $categories = Category::whereIn('id', $fields['categories'])->get();
+
+            $post->categories()->sync($categories);
+
+            return new PostResource($post);
+        } else {
+            return response([
+                'message' => 'you cannot update a post that is not yours',
+            ], 403);
+        }
+
     }
 
     /**
@@ -49,8 +79,15 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        $post->delete();
+        if (auth()->user()->getAuthIdentifier() == $post->user_id) {
+            $post->delete();
 
-        return response(null, \Illuminate\Http\Response::HTTP_NO_CONTENT);
+            return response(null, ResponseAlias::HTTP_NO_CONTENT);
+        } else {
+            return response([
+                'message' => 'you cannot delete a post that is not yours',
+            ], 403);
+        }
+
     }
 }
